refactor(comingsoon): rotating text component
